% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ena_2023.R
\name{download_ena_2023}
\alias{download_ena_2023}
\title{Download Modules of the National Agricultural Survey 2023 (INEI - Peru)}
\usage{
download_ena_2023(modulo = "all", dir_output = "data_ena_peru")
}
\arguments{
\item{modulo}{A character vector specifying which modules to download.
Use \code{"all"} (default) to download all modules, or specify one or more module names
such as \code{"modulo1830"}, \code{"modulo1831"}, etc. Use \code{list_ena_2023_modules()} to see available modules.}

\item{dir_output}{A character string specifying the directory where the files
will be saved and extracted. Defaults to \code{"data_ena_peru"}.}
}
\value{
A character string with the absolute path of the output directory where the files were extracted.
}
\description{
Downloads one, several, or all data modules from the
National Agricultural Survey 2023 (Encuesta Nacional Agropecuaria ENA-2023),
conducted by the National Institute of Statistics and Informatics (INEI) of Peru.
}
\details{
Module descriptions are based on INEI documentation.
}
