% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ena_2023_modules.R
\name{list_ena_2023_modules}
\alias{list_ena_2023_modules}
\title{List of Available ENA 2023 Modules (Spanish and English)}
\usage{
list_ena_2023_modules()
}
\value{
A \code{tibble} with 4 columns:
\describe{
\item{module_code}{Official numeric code of the module as defined by INEI.}
\item{module_tag}{Internal name or tag used in the package, useful for download functions.}
\item{modulo_es}{Spanish title of the module, as published by INEI.}
\item{module_en}{English translation of the module title, useful for non-Spanish-speaking users.}
}
}
\description{
Returns the codes, internal tags, and titles of all available modules from the
2023 National Agricultural Survey (Encuesta Nacional Agropecuaria - ENA) conducted
by the National Institute of Statistics and Informatics (INEI) of Peru.
The module titles are provided in both Spanish and English to facilitate access for native users
and international audiences.
}
\examples{
list_ena_2023_modules()

}
